renamed folders
removed the scroll

add index.js

transferred styles to components

added absolute paths
