+++ src/App.js
@@ -1,13 +1,13 @@
 import React from "react";
 import video from "./public/header.mp4";
 import img from "./public/cut_the_rope.png";
-import Weather from "./Container/screens/Weather";
+import Weather from "./container/screens/weather/Weather";
 import "./App.css";
 
 function App() {
   return (
-    <div className="App">
-      <video className="videoTag" autoPlay loop muted width="100%" poster={img}>
+    <div>
+      <video autoPlay loop muted width="100%" poster={img}>
         <source src={video} type="video/mp4" />
       </video>
       <Weather />
