+++ src/components/comment/index.js
@@ -1,0 +1,3 @@
+import Comments from "./Comment";
+
+export default Comments;
