+++ src/components/header/Header.css
@@ -1,0 +1,7 @@
+.header {
+  background-color: hotpink;
+  opacity: 0.5;
+  height: 10%;
+  border-radius: 15px 15px 0px 0px;
+  align-items: center;
+}
