+++ src/components/header/Header.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { Col, Row } from "antd";
+import Search from "../search/Search";
+import "./Header.css";
+
+function Header() {
+  return (
+    <Row className="header">
+      <Col span={8} />
+      <Col span={8} className="search">
+        <Search />
+      </Col>
+      <Col span={8} />
+    </Row>
+  );
+}
+
+export default Header;
