+++ src/components/headerWeather/HeaderWeather.css
@@ -1,0 +1,6 @@
+.headerWeather {
+  height: 10%;
+  background-color: darkgrey;
+  align-items: flex-end;
+  padding-left: 3%;
+}
