+++ src/components/headerWeather/HeaderWeather.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { Row } from "antd";
+import "./HeaderWeather.css";
+
+function HeaderWeather() {
+  return (
+    <Row className="headerWeather">
+      <h1>Москва</h1>
+    </Row>
+  );
+}
+
+export default HeaderWeather;
