+++ src/components/headerWeather/index.js
@@ -1,0 +1,3 @@
+import HeaderWeather from "./HeaderWeather";
+
+export default HeaderWeather;
