+++ src/components/weatherWidget/WidgetsStyle.css
