+++ src/components/weatherWidget/index.js
@@ -1,0 +1,3 @@
+import WeatherWidget from "./WeatherWidget";
+
+export default WeatherWidget;
