+++ src/container/screens/weather/Weather.css
@@ -1,0 +1,34 @@
+.main {
+  position: absolute;
+  top: 5%;
+  left: 10%;
+  height: 90%;
+  width: 80%;
+  background-color: white;
+  opacity: 0.9;
+  border-radius: 15px;
+}
+.content {
+  height: 80%;
+}
+
+.info {
+  height: 100%;
+}
+.clothes {
+  background-color: darkgrey;
+  opacity: 0.5;
+}
+.clothesPhotos {
+  height: 80%;
+}
+.clothesComment {
+  height: 20%;
+  align-items: flex-end;
+}
+.footer {
+  background-color: pink;
+  opacity: 0.5;
+  height: 10%;
+  border-radius: 0px 0px 15px 15px;
+}
